Meta cleaning (#331)
* policy

* cleanup

* removed unneeded

* bad-notebook

* mixture

* windows-via-actions

* removed appveyor
