--- .appveyor.yml
@@ -1,24 +1,0 @@
-build: false
-
-environment:
-  global:
-      REQUIREMENTS: "numpy scipy scikit-learn pytest flake8"
-
-  matrix:
-    - PYTHON_VERSION: 3.6
-      MINICONDA: C:\Miniconda3-x64
-
-
-init:
-  - "ECHO %PYTHON_VERSION% %MINICONDA%"
-
-install:
-  - "set PATH=%MINICONDA%;%MINICONDA%\\Scripts;%PATH%"
-  - conda config --set always_yes yes --set changeps1 no
-  - conda create -n sklego-env -c %REQUIREMENTS% python=%PYTHON_VERSION%
-  - activate sklego-env
-  - pip install -e .[dev]
-
-test_script:
-  - activate sklego-env
-  - pytest tests/
