+++ .github/workflows/test-linux.yml
@@ -1,4 +1,4 @@
-name: Python package
+name: Python Package Linux
 
 on:
   push:
