+++ sklego/mixture/__init__.py
@@ -1,0 +1,11 @@
+__all__ = [
+    "GMMClassifier",
+    "BayesianGMMClassifier",
+    "GMMOutlierDetector",
+    "BayesianGMMOutlierDetector"
+]
+
+from .gmm_classifier import GMMClassifier
+from .bayesian_gmm_classifier import BayesianGMMClassifier
+from .gmm_outlier_detector import GMMOutlierDetector
+from .bayesian_gmm_detector import BayesianGMMOutlierDetector
